+++ packages/code/CHANGELOG.md
@@ -4,12 +4,12 @@
 
 ### Major Changes
 
-- 717e2cf: Updates Code to allign with visual brand update
+- ab1fd9e: Updates Code to allign with visual brand update
   - Removes fixed width when Language picker is shown (`width` was formerly set to a static `700px`)
 
 ### Patch Changes
 
-- Updated dependencies [717e2cf]
+- Updated dependencies [ab1fd9e]
   - @leafygreen-ui/icon-button@10.0.0
 
 ## 9.5.0