+++ packages/modal/CHANGELOG.md
@@ -1,10 +1,14 @@
 # @leafygreen-ui/modal
 
-## 7.0.1
+## 8.0.0
+
+### Major Changes
+
+- ab1fd9e: Updates modal and confirmation modal styles in line with visual brand refresh. Add prop closeIconColor.
 
 ### Patch Changes
 
-- Updated dependencies [717e2cf]
+- Updated dependencies [ab1fd9e]
   - @leafygreen-ui/icon-button@10.0.0
 
 ## 7.0.0