+++ packages/tooltip/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Major Changes
 
-- 717e2cf: - Updates tooltip for visual brand refresh.
+- ab1fd9e: - Updates tooltip for visual brand refresh.
   - Updates darkMode for Tooltip
   - Using `usePortal={false}` will set the tooltip to the width of the content using `max-content` (up to a `max-width` of 256px). You can override this behavior using `portalClassName`. (Override the max-width using `className` on the Tooltip)